Update 2:
- Objects can now be spawned, the prefab must have the"objectData"
script added too it, the information in the script is the options for
generation. the min and max percentage values is how far from the center
of the path thigs can be: min 0, max 100 lets it generate anywhere on
the path. min 50, max 60 means it can only generate past the 50% mark of
the Side-> centre area and the furthest it can generate is 60% of that
area...have fun with that